minor: Add support for files requiring virus scan confirmation
The download was not working for large files (or particular file structures
like zip of zips) where Google Drive was unable to scan the file for viruses.

Multiple MRs addressed the same issue parsing the first response content:
https://github.com/ndrplz/google-drive-downloader/pull/28
https://github.com/ndrplz/google-drive-downloader/pull/32/

However, it seems that simply replacing the GET with a POST already cuts it.

Possible explanations from ChatGPT:
- Google Drive treats POST as an intentional download request
- Google might have a different backend handling for POST downloads
- The confirm=True parameter might act as a global override

In any case, it seems to work in both cases now.